@@ -6,6 +6,7 @@ import java.awt.*;
 import java.awt.event.*;
 import java.awt.image.BufferedImage;
 import java.io.IOException;
+import java.net.MalformedURLException;
 import java.net.URL;
 
 public class ShowEntry extends JPanel {
@@ -16,25 +17,33 @@ public class ShowEntry extends JPanel {
     private JLabel iconLabel;
     private JButton selectButton;
 
+    private boolean textCropped = false;
     private int number;
 
-    private static int width = 200;
+    private static final int WIDTH = 200;
+    private static final int MAX_STRING_LENGTH = 25;
 
-    public ShowEntry(String text, String iconPath, final int number, final GUI gui) {
+    public ShowEntry(String text, String iconPath, final int number, final GUI gui) throws IOException {
         super();
 
-        //LABEL ICON
         textArea = new JTextArea(text);
-        textArea.setLineWrap(true);
+
+        //LABEL ICON
+        if(text.length() > MAX_STRING_LENGTH){
+            textCropped = true;
+            textArea.setText(text.replaceAll("(?<=^.{"+MAX_STRING_LENGTH+"}).*", "..."));
+        }
+
+        setToolTipText(text);
+
+        textArea.setEditable(false);
         selectButton = new JButton("Select Show");
 
         iconLabel = new JLabel();
-        try {
-            thumbnail = ImageIO.read(new URL(iconPath));
-        } catch (IOException ignored) {
-        }
-        int scaledHeight = (int) ((double) thumbnail.getHeight() / (double) thumbnail.getWidth() * width);
-        Image scaled = thumbnail.getScaledInstance(width, scaledHeight, Image.SCALE_DEFAULT);
+        thumbnail = ImageIO.read(new URL(iconPath));
+
+        int scaledHeight = (int) ((double) thumbnail.getHeight() / (double) thumbnail.getWidth() * WIDTH);
+        Image scaled = thumbnail.getScaledInstance(WIDTH, scaledHeight, Image.SCALE_DEFAULT);
         icon = new ImageIcon(scaled);
         iconLabel.setIcon(icon);
 
